Search the Cal-Adapt raster series data catalog

Search the Cal-Adapt raster series data catalog

ca_catalog_search(x, keep_together = FALSE, quiet = FALSE)

Arguments

x
text to search for

keep_together
treat x as a phrase, logical

quiet
suppress messages

Value

A tibble with information about the found raster series

Details

This function can be used to search the local copy of the Cal-Adapt raster series data catalog, and view the properties of the matching results. Searched fields include the dataset name and slug. If keep_together = TRUE, the search text will be treated as a phrase, otherwise the words in x will be searched for separately. Records have to match all terms to be returned.

For an online search tool, click the 'Filters' button on https://api.cal-adapt.org/api/series/.

See also

Examples

if (FALSE) {
## Search for a slug
ca_catalog_search("pr_day_gridmet")

## Search for keywords
ca_catalog_search("evapotranspiration year")

## Search for phrase
ca_catalog_search("Livneh VIC", keep_together = TRUE)
}