Add point location(s) to a Cal-Adapt API request

Specifies point location(s) a Cal-Adapt API call should retrieve

ca_loc_pt(x = ca_apireq(), coords, id = NULL)

Arguments

x

A Cal-Adapt API request

coords

A two-column matrix or data frame

id

Unique id values for each point

Details

coords should be a two-column matrix or data frame with the first column containing the x (longitude) values of the points of interest, and the second column containing the y (latitude) values. Projected coordinates can not be used with this function (but see ca_loc_sf).

id should be vector that uniquely identify the points. If omitted, row numbers will be used.

See also