Get the LOCA grid cells as a sf object

Get the geometry of the LOCA grid cells as a sf polygon object

Usage

ca_locagrid_geom(quiet = FALSE)

Arguments

quiet

Suppress messages

Value

A simple feature data frame

Details

This retrieves the geometry of the LOCA grid as a vector (polygon) layer. The cells in this grid represent the pixels for all the LOCA downscaled raster series on Cal-Adapt. A copy of the layer will be saved in the cache folder so it won't have to be downloaded more than once.

See also