Create an index for a list of stars rasters

Create an index for a list of stars rasters

Usage

ca_stars_index(x)

Arguments

x

A list of stars rasters

Value

A tibble with the properties of elements of x. There will be one row for each element of x. Columns include cvar, scenario, gcm, period, slug, livneh, start, end, rows, and cols.

Details

When you download rasters from Cal-Adapt using ca_getrst_stars ca_stars_index generates an index of the properties of the elements of a list of stars rasters to help identify which stars rasters contain which climate model data.

See also

Examples

if (FALSE) { # \dontrun{
## Download 5 years of daily max and min temp for Merced County as rasters
mercd_cap <- ca_loc_aoipreset(type = "counties", idfld = "fips", idval = "06047") %>%
  ca_gcm(gcms[1:2]) %>%
  ca_period("day") %>%
  ca_cvar(c("tasmin", "tasmax")) %>%
  ca_scenario("rcp45") %>%
  ca_years(start = 2060, end = 2065)

mercd_stars_lst <- mercd_cap %>%
  ca_getrst_stars(out_dir = ".") %>%
  ca_read_stars()

## Create an index tibble to see the climate model in each stars raster
mercd_stars_tbl <- mercd_stars_lst %>%
  ca_starslist_index()
} # }