Mosaic stars objects into a seamless array for large areas

Mosaic stars objects into a seamless array for large areas

Usage

ca_stars_mosaic(
  stars_lst,
  index_tbl = NULL,
  geom_mask = NULL,
  combine_6d = FALSE,
  quiet = FALSE
)

Arguments

stars_lst

A list of stars rasters

index_tbl

A tibble of metadata for stars_lst

geom_mask

A sf or sfc polygon object to crop the mosaiced raster

combine_6d

Combine multiple 3D rasters into one 6D raster

quiet

Suppress messages

Value

If combine_6d = FALSE, a list of 3D rasters will be returned (x, y and date/year). If combine_6d = TRUE, a 6D stars raster will be returned (x, y, date/year, scenario, GCM, climate variable).

Details

stars_lst is a list of stars objects downloaded by ca_getrst_stars and turned into a list by ca_stars_read. Note that both of these functions must be run with `sidecar = TRUE` (the default).

combine_6d = TRUE will return the mosaic as a six-dimensional raster. This further requires that the raster was originally downloaded using an API request that specified the dataset by scenario, GCM, and climate variable (i.e., not by a slug).

See also

Examples

if (FALSE) { # \dontrun{

} # }