Global climate models

Global Climate Models available through the Cal-Adapt API

Usage

data(gcms)

Format

An character vector with the names of 10 GCMs (abbreviated)

Source

https://berkeley-gif.github.io/caladapt-docs/data-catalog.html#global-climate-models-gcm

Details

The following GCMs have been selected by California state agencies as priority models for Fourth Assessment Research and are available thru the Cal-Adapt API. The first four have been identified as the priority GCMs representing (see 4 priority models)

HadGEM2-ES: Met Office Hadley Centre and Instituto Nacional de Pesquisas Espaciais

CNRM-CM5: Centre National de Recherches Météorologiques/ Centre Européen de Recherche et Formation Avancée en Calcul Scientifique

CanESM2: Canadian Centre for Climate Modelling and Analysis

MIROC5: Atmosphere and Ocean Research Institute (The University of Tokyo), National Institute for Environmental Studies, and Japan Agency for Marine-Earth Science and Technology

ACCESS1-0: Commonwealth Scientific and Industrial Research Organization (CSIRO) and Bureau of Meteorology (BOM) Australia

CCSM4: University of Miami - RSMAS

CESM1-BGC: Community Earth System Model Contributors

CMCC-CMS: Centro Euro-Mediterraneo per I Cambiamenti Climatici

GFDL-CM3: NOAA Geophysical Fluid Dynamics Laboratory

HadGEM2-CC: Met Office Hadley Centre