Internal function used by ca_apicalls() to handle issues discovered.

This is mostly to keep the code clean. If pf=TRUE, they'll just be logged in genl_msgs, gval_msgs, grst_msgs. stop_msg is the message to show if the problem is a deal-breaker gval_msgs, grst_msgs are the messages to show if the context is getvals or getrst (these are sometimes different) If neither gval_msgs or grst_msgs are passed, the error will be logged (pf=TRUE) or code stopped (pf=FALSE) If pf = FALSE and stop_msg has value, the stop() will be called

stopwarn(pf, check_for, stop_msg = NULL, gval_msg = NULL, grst_msg = NULL)