Manage package settings

View and manage package settings

ca_settings(console_colors = NA, date_slice = NA, quiet = FALSE)

Arguments

console_colors

The name of a preset, or list

date_slice

Whether to use date slicing on the Cal-Adapt API when available

quiet

Suppress messages, logical

Details

console_colors controls the color of text printed at the console. You can pass the name of a preset or a named list of color functions (i.e., from crayon package). Up to six styles are recognized, see example below.

date_slice determines whether or not date slicing via URL construction should be used for those Cal-Adapt datasets that support it. This is generally a good idea, but can be set to FALSE for trouble-shooting.

See also

Examples

if (FALSE) {
ca_settings(console_colors = list(ca_accent1 = crayon::blue$bold,
                                  ca_accent2 = crayon::magenta,
                                  ca_accent3 = crayon::red$bold,
                                  ca_accent4 = crayon::red,
                                  ca_message = crayon::silver,
                                  ca_success = crayon::green$bold))
}