Create a six-dimensional stars object for modeled climate data

Create a six-dimensional stars object for modeled climate data

Usage

ca_stars_6d(stars_lst, index_tbl = NULL)

Arguments

stars_lst

A list of stars rasters

index_tbl

A tibble of metadata for stars_lst

Value

A six-dimensional stars object with dimensions x, y, scenario, gcm, date, and cvar

Details

stars_lst is a list of stars objects downloaded by ca_getrst_stars and turned into a list by ca_stars_read. Note that both of these functions must use `sidecar = TRUE`.

Creating a six-dimensional stars array of projected climate data may be useful for writing more compact expressions for analysis. Six-dimensional arrays can only be constructed if the API request specifed the GCM, scenario, and climate variable. Rasters retrieved using an API request that specified the dataset by the name of the slug can not be turned into a 6D arrays. Another requirement is that all the rasters have the same location / extent.

See also

Examples

if (FALSE) { # \dontrun{

} # }